adjust spark imports/exports
