+++ splink/databricks/enable_splink.py
@@ -1,4 +1,4 @@
-from splink.spark.jar_location import similarity_jar_location
+from splink.spark.spark_helpers.jar_location import similarity_jar_location
 
 
 def enable_splink(spark):
