+++ splink/spark/spark_comparison_level_library.py
@@ -1,121 +1,16 @@
 from ..comparison_level_composition import and_, not_, or_  # noqa: F401
-from ..comparison_level_library import (
-    ArrayIntersectLevelBase,
-    ColumnsReversedLevelBase,
-    DateDiffLevelBase,
-    DistanceFunctionLevelBase,
-    DistanceInKMLevelBase,
-    ElseLevelBase,
-    ExactMatchLevelBase,
-    JaccardLevelBase,
-    JaroLevelBase,
-    JaroWinklerLevelBase,
-    LevenshteinLevelBase,
-    NullLevelBase,
-    PercentageDifferenceLevelBase,
+from .spark_helpers.spark_comparison_imports import (  # noqa: F401
+    array_intersect_level,
+    columns_reversed_level,
+    datediff_level,
+    distance_function_level,
+    distance_in_km_level,
+    else_level,
+    exact_match_level,
+    jaccard_level,
+    jaro_level,
+    jaro_winkler_level,
+    levenshtein_level,
+    null_level,
+    percentage_difference_level,
 )
-from .spark_base import (
-    SparkBase,
-)
-
-
-class SparkComparisonProperties(SparkBase):
-    @property
-    def _exact_match_level(self):
-        return exact_match_level
-
-    @property
-    def _null_level(self):
-        return null_level
-
-    @property
-    def _else_level(self):
-        return else_level
-
-    @property
-    def _datediff_level(self):
-        return datediff_level
-
-    @property
-    def _array_intersect_level(self):
-        return array_intersect_level
-
-    @property
-    def _columns_reversed_level(self):
-        return columns_reversed_level
-
-    @property
-    def _distance_in_km_level(self):
-        return distance_in_km_level
-
-    @property
-    def _distance_function_level(self):
-        return distance_function_level
-
-    @property
-    def _levenshtein_level(self):
-        return levenshtein_level
-
-    @property
-    def _jaro_level(self):
-        return jaro_level
-
-    @property
-    def _jaro_winkler_level(self):
-        return jaro_winkler_level
-
-    @property
-    def _jaccard_level(self):
-        return jaccard_level
-
-
-class null_level(SparkBase, NullLevelBase):
-    pass
-
-
-class exact_match_level(SparkBase, ExactMatchLevelBase):
-    pass
-
-
-class else_level(SparkBase, ElseLevelBase):
-    pass
-
-
-class columns_reversed_level(SparkBase, ColumnsReversedLevelBase):
-    pass
-
-
-class distance_function_level(SparkBase, DistanceFunctionLevelBase):
-    pass
-
-
-class levenshtein_level(SparkBase, LevenshteinLevelBase):
-    pass
-
-
-class jaro_level(SparkBase, JaroLevelBase):
-    pass
-
-
-class jaro_winkler_level(SparkBase, JaroWinklerLevelBase):
-    pass
-
-
-class jaccard_level(SparkBase, JaccardLevelBase):
-    pass
-
-
-class array_intersect_level(SparkBase, ArrayIntersectLevelBase):
-    pass
-
-
-class percentage_difference_level(SparkBase, PercentageDifferenceLevelBase):
-    pass
-
-
-class distance_in_km_level(SparkBase, DistanceInKMLevelBase):
-    pass
-
-
-class datediff_level(SparkBase, DateDiffLevelBase):
-    pass
