+++ splink/spark/spark_comparison_library.py
@@ -1,72 +1,11 @@
-from ..comparison_library import (
-    ArrayIntersectAtSizesComparisonBase,
-    DateDiffAtThresholdsComparisonBase,
-    DistanceFunctionAtThresholdsComparisonBase,
-    DistanceInKMAtThresholdsComparisonBase,
-    ExactMatchBase,
-    JaccardAtThresholdsComparisonBase,
-    JaroAtThresholdsComparisonBase,
-    JaroWinklerAtThresholdsComparisonBase,
-    LevenshteinAtThresholdsComparisonBase,
+from .spark_helpers.spark_comparison_imports import (  # noqa: F401
+    array_intersect_at_sizes,
+    datediff_at_thresholds,
+    distance_function_at_thresholds,
+    distance_in_km_at_thresholds,
+    exact_match,
+    jaccard_at_thresholds,
+    jaro_at_thresholds,
+    jaro_winkler_at_thresholds,
+    levenshtein_at_thresholds,
 )
-from .spark_comparison_level_library import SparkComparisonProperties
-
-
-class exact_match(SparkComparisonProperties, ExactMatchBase):
-    pass
-
-
-class distance_function_at_thresholds(
-    SparkComparisonProperties, DistanceFunctionAtThresholdsComparisonBase
-):
-    @property
-    def _distance_level(self):
-        return self._distance_function_level
-
-
-class levenshtein_at_thresholds(
-    SparkComparisonProperties, LevenshteinAtThresholdsComparisonBase
-):
-    @property
-    def _distance_level(self):
-        return self._levenshtein_level
-
-
-class jaro_at_thresholds(SparkComparisonProperties, JaroAtThresholdsComparisonBase):
-    @property
-    def _distance_level(self):
-        return self._jaro_level
-
-
-class jaro_winkler_at_thresholds(
-    SparkComparisonProperties, JaroWinklerAtThresholdsComparisonBase
-):
-    @property
-    def _distance_level(self):
-        return self._jaro_winkler_level
-
-
-class jaccard_at_thresholds(
-    SparkComparisonProperties, JaccardAtThresholdsComparisonBase
-):
-    @property
-    def _distance_level(self):
-        return self._jaccard_level
-
-
-class array_intersect_at_sizes(
-    SparkComparisonProperties, ArrayIntersectAtSizesComparisonBase
-):
-    pass
-
-
-class datediff_at_thresholds(
-    SparkComparisonProperties, DateDiffAtThresholdsComparisonBase
-):
-    pass
-
-
-class distance_in_km_at_thresholds(
-    SparkComparisonProperties, DistanceInKMAtThresholdsComparisonBase
-):
-    pass
