+++ splink/spark/spark_linker.py
@@ -18,7 +18,7 @@ from ..linker import Linker
 from ..misc import ensure_is_list, major_minor_version_greater_equal_than
 from ..splink_dataframe import SplinkDataFrame
 from ..term_frequencies import colname_to_tf_tablename
-from .custom_spark_dialect import Dialect
+from .spark_helpers.custom_spark_dialect import Dialect
 
 logger = logging.getLogger(__name__)
